Добавляет экспорт, импорт функций
